Automate testing against 'gold'
For each sample file, create expected output file (gold),
which is saved in git together with the sample file.
Then during testing  is expected to produce exact
match to the 'gold'.

Of course, input and '.conf' file has to be exactly the same.

While limited, this can prevent changes in output going unnoticed.